requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-CLI==0.4.0
googledrivedownloader==0.4
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
opencv-python-headless==4.2.0.32
Pillow==8.2.0
PySimpleGUI==4.40.0
requests==2.25.1
tk==0.1.0
urllib3==1.26.4
Werkzeug==1.0.1
